cc-pipeline/pyproject.toml
[project]
name = "cc-pipeline"
version = "0.1.0"
description = "Module A: Crawl + Clean pipeline for RAG knowledge resources."
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.2.0",
  "pyyaml>=6.0.0",
  "python-dotenv>=1.0.0",
  "pypdf>=5.0.0",
  "tenacity>=8.2.0",
  "orjson>=3.10.0",
  "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
]

[project.scripts]
rag-cc = "cc_pipeline.cli:main"

[build-system]
requires = ["setuptools>=69.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
